requirements.txt
streamlit
pyfinance
dill
plotly-express
numpy==1.26.4
yfinance